HUGE CHANGES
Added full support of database. Reworked the models. Created separate tables for player, legends, badges and trackers. This was done using ForeignKeys and should improve the whole read/write process with database

+++ mozambique-api/src/main/java/com/skullper/mozambique_api/response/player/Realtime.kt
@@ -7,7 +7,7 @@ data class Realtime(
     val canJoin: Int,
     val currentState: String,
     val currentStateAsText: String,
-    val currentStateSinceTimestamp: Int,
+    val currentStateSinceTimestamp: Long,
     val isInGame: Int,
     val isOnline: Int,
     val lobbyState: String,
